backend/requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==40.0.2
Django==4.2.1
django-cors-headers==3.14.0
django-secured-fields==0.4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
environ==1.0
icalendar==5.0.5
Pillow==9.5.0
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
